feat : 이미지 파일 유효성 검사 훅
